refactor: Use File objects for ClothingItem images
This commit replaces the base64 string representation of images with
File objects in the ClothingItem class for space considerations.  This change allows us to get
rid of the somewhat complicated "base64ToBlob" function.  It also
introduces caching of the base64 representation so we don't have to
regenerate the base64 string on every paint.